Separate collecting concepts from resolving a type
TypeRegistry had grown past the size a file is meant to stay under by
answering two questions at once: what name to write for a type, and what
the application's concepts are. The second is a registry with rules of
its own - what counts as a concept, which declaration wins when two share
a name, how a personal-data mark and a validation rule arriving later
still find the concept they belong to - and none of that has anything to
do with unwrapping a collection or reporting what a name loses.

ConceptRegistry now owns the concepts, and TypeRegistry keeps type
resolution and the two things that fall out of writing a name. The
enumeration and ConceptAs branches that were repeated between naming a
type and walking what it carries collapse into one TryRegister, which is
the same decision stated once.

Behavior is unchanged: the document generated for a real application is
byte-identical before and after, and all 973 specs pass.
